+++ src/games/colonizers/phase/index.ts
@@ -1,0 +1,5 @@
+export enum Phase {
+  Place = 'Place', // Initial settlement and road placement
+  Game = 'Game',
+  Robber = 'Robber', // Move robber
+}
